Split admin metrics server helpers
