+++ tests/admin-metrics-server-architecture.test.ts
@@ -5,20 +5,52 @@ import test from 'node:test';
 
 const root = process.cwd();
 const metricsPath = join(root, 'frontend/server/admin-metrics.ts');
+const metricsDir = join(root, 'frontend/server/admin-metrics');
 const helpersPath = join(root, 'frontend/server/admin-metrics/admin-metrics-helpers.ts');
+const mainPath = join(metricsDir, 'admin-metrics-main.ts');
+const comparisonPath = join(metricsDir, 'admin-metrics-comparison.ts');
+const healthPath = join(metricsDir, 'admin-health.ts');
+const engineUsagePath = join(metricsDir, 'admin-metrics-engine-usage.ts');
+const shapingPath = join(metricsDir, 'admin-metrics-shaping.ts');
+const focusedModules = [
+  'admin-health.ts',
+  'admin-metrics-comparison.ts',
+  'admin-metrics-engine-usage.ts',
+  'admin-metrics-helpers.ts',
+  'admin-metrics-main.ts',
+  'admin-metrics-shaping.ts',
+];
 
 const metricsSource = readFileSync(metricsPath, 'utf8');
 const helpersSource = readFileSync(helpersPath, 'utf8');
+const mainSource = readFileSync(mainPath, 'utf8');
+const comparisonSource = readFileSync(comparisonPath, 'utf8');
+const healthSource = readFileSync(healthPath, 'utf8');
+const engineUsageSource = readFileSync(engineUsagePath, 'utf8');
+const shapingSource = readFileSync(shapingPath, 'utf8');
 
-test('admin metrics delegates shared row helpers and range logic', () => {
+test('admin metrics public module delegates to focused server modules', () => {
   assert.ok(existsSync(helpersPath), 'admin metrics helpers should live in a focused server module');
+  assert.match(metricsSource, /from '@\/server\/admin-metrics\/admin-metrics-main'/);
+  assert.match(metricsSource, /from '@\/server\/admin-metrics\/admin-metrics-comparison'/);
+  assert.match(metricsSource, /from '@\/server\/admin-metrics\/admin-health'/);
+  assert.match(metricsSource, /from '@\/server\/admin-metrics\/admin-metrics-engine-usage'/);
   assert.match(metricsSource, /from '@\/server\/admin-metrics\/admin-metrics-helpers'/);
-  assert.match(metricsSource, /fetchAdminMetrics\(/);
-  assert.match(metricsSource, /fetchAdminMetricsComparison\(/);
-  assert.match(metricsSource, /fetchAdminHealth\(/);
+
+  const lineCount = metricsSource.split('\n').length;
+  assert.ok(lineCount <= 40, `admin-metrics.ts should stay a thin facade, got ${lineCount} lines`);
 });
 
-test('admin metrics fetcher does not regain shared helper ownership', () => {
+test('admin metrics modules stay focused and under the large-file threshold', () => {
+  for (const moduleName of focusedModules) {
+    const modulePath = join(metricsDir, moduleName);
+    assert.ok(existsSync(modulePath), `${moduleName} should exist under admin-metrics`);
+    const lineCount = readFileSync(modulePath, 'utf8').split('\n').length;
+    assert.ok(lineCount <= 500, `${moduleName} should stay below 500 lines, got ${lineCount}`);
+  }
+});
+
+test('admin metrics facade does not regain query or helper ownership', () => {
   for (const pattern of [
     /type CountRow =/,
     /type AmountRow =/,
@@ -35,12 +67,12 @@ test('admin metrics fetcher does not regain shared helper ownership', () => {
     /function fillAmountSeries\(/,
     /function buildEmptyMetrics\(/,
     /function mapEngineUsage\(/,
+    /safeQuery</,
+    /SELECT\s+/,
+    /WITH\s+/,
   ]) {
     assert.doesNotMatch(metricsSource, pattern);
   }
-
-  const lineCount = metricsSource.split('\n').length;
-  assert.ok(lineCount <= 820, `admin-metrics.ts should stay below 820 lines after helper extraction, got ${lineCount}`);
 });
 
 test('admin metrics helper module exposes the expected contract', () => {
@@ -73,4 +105,15 @@ test('admin metrics helper module exposes the expected contract', () => {
   ]) {
     assert.match(helpersSource, new RegExp(`export (type |const |function |async function )${exportName}`));
   }
+});
+
+test('admin metrics focused modules expose the expected route contracts', () => {
+  assert.match(mainSource, /export async function fetchAdminMetrics/);
+  assert.match(comparisonSource, /export async function fetchAdminMetricsComparison/);
+  assert.match(healthSource, /export async function fetchAdminHealth/);
+  assert.match(engineUsageSource, /export async function fetchEngineUsageMetrics/);
+  assert.match(engineUsageSource, /export async function loadEngineUsageRows/);
+  assert.match(shapingSource, /export function buildFunnelMetrics/);
+  assert.match(shapingSource, /export function buildBehaviorMetrics/);
+  assert.match(shapingSource, /export function buildHealthMetrics/);
 });
